requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.3
django-crispy-forms==1.9.0
django-extensions==2.2.8
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
